+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Files: fix external link to folder in folder panel
+
 23-06-2025: 25.0.32
 - LDAPServer: Fix primary key error